Guard file IO by skipping global PRIME
Drop the global PRIME definition in favour of a prime function, to
avoid unnecessary file IO on module import. This requires some renaming
of "prime" arguments.

Consolidate hashlib imports by not importing sha256 separately.

Create a new IntegerHasher class that implements the _integer_hash
method. This gets inherited by SRPPeer (multiple inheritance with ABC),
but it is split out in anticipation of using it in m38.